decrating: collapse dual implementation for storage (preserve shim logic)
Eliminates the dual-path pattern for the storage subsystem:
- Deletes the stale in-tree `storage.rs`.
- Promotes `storage_micro.rs` to canonical `storage.rs`.
- The shim's added logic is preserved:
  - `FileStorage` wrapper struct providing the historical
    `base_path() -> &PathBuf` accessor (the microcrate exposes `path()`
    instead), keeping the old `shipper::storage` API stable.
  - `build_storage_backend` short-circuits for `StorageType::File` to
    use the wrapped `FileStorage`, delegating to the microcrate only
    for cloud backends.
  - `config_from_env` accepts extra aliases: `local` (for `file`),
    `gs` (for `gcs`), and `blob` (for `azure`).
- Removes `micro-storage` feature from shipper and shipper-cli
  Cargo.toml, makes `shipper-storage` a non-optional dep, drops the
  cross-reference from `micro-config` and `micro-all`.
