+++ crates/shipper/Cargo.toml
@@ -20,7 +20,7 @@ shipper-git = { path = "../shipper-git", optional = true }
 shipper-events = { path = "../shipper-events" }
 shipper-execution-core = { path = "../shipper-execution-core" }
 shipper-environment = { path = "../shipper-environment" }
-shipper-storage = { path = "../shipper-storage", optional = true }
+shipper-storage = { path = "../shipper-storage" }
 shipper-cargo = { path = "../shipper-cargo", optional = true }
 shipper-engine-parallel = { path = "../shipper-engine-parallel", optional = true }
 shipper-plan = { path = "../shipper-plan", optional = true }
@@ -63,14 +63,13 @@ sha2 = "0.11"
 [features]
 micro-git = ["dep:shipper-git"]
 micro-encrypt = []
-micro-storage = ["dep:shipper-storage"]
 micro-cargo = ["dep:shipper-cargo"]
 micro-plan = ["dep:shipper-plan", "micro-types"]
 micro-webhook = []
 micro-types = ["micro-webhook", "micro-encrypt", "micro-config"]
-micro-config = ["micro-types", "micro-storage"]
+micro-config = ["micro-types"]
 micro-parallel = ["micro-types", "micro-plan", "dep:shipper-engine-parallel"]
-micro-all = ["micro-git", "micro-encrypt", "micro-storage", "micro-cargo", "micro-plan", "micro-webhook", "micro-types", "micro-config", "micro-parallel"]
+micro-all = ["micro-git", "micro-encrypt", "micro-cargo", "micro-plan", "micro-webhook", "micro-types", "micro-config", "micro-parallel"]
 
 [lints]
 workspace = true
